requirements.txt
appdirs==1.4.4
arrow==0.17.0
atomicwrites==1.4.0
attrs==20.3.0
awscli==1.18.159
bcrypt==3.2.0
boto3==1.15.18
botocore==1.18.18
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
distlib==0.3.1
dnspython==2.0.0
docutils==0.15.2
dominate==2.5.2
email-validator==1.1.1
filelock==3.0.12
firebase==3.0.1
firebase-admin==4.4.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-S3==0.3.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gcloud==0.18.3
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.2
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jws==0.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
oauth2client==4.1.2
packaging==20.4
pdfkit==0.6.1
pipenv==2020.8.13
pluggy==0.13.1
protobuf==3.13.0
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
pyparsing==2.4.7
Pyrebase==3.0.27
Pyrebase4==4.4.1
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jwt==2.0.1
python-pdf==0.37
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.5
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.10
uuid==1.30
virtualenv==20.0.31
virtualenv-clone==0.5.4
visitor==0.1.3
Werkzeug==1.0.1
whitenoise==5.2.0
wkhtmltopdf-pack-ng==0.12.3.0
WTForms==2.3.3
zipp==3.1.0
